// apps/web/src/lib/api/types.gen.ts
export type GetYearRecapResponses = {
    /**
     * Year-to-date statistics about emergency response
     */
    200: {
        /**
         * Top 5 days with the most incidents in the selected year.
         */
        topIncidentDays: Array<{
            /**
             * Date in YYYY-MM-DD format.
             */
            date: string;
            /**
             * Number of incidents on that date.
             */
            count: number;
        }>;
        /**
         * Top 5 stations with the most dispatches in the selected year.
         */
        topDispatchedStations: Array<{
            /**
             * Station name.
             */
            name: string;
            /**
             * Number of dispatches.
             */
            count: number;
        }>;
        /**
         * Top 5 vehicles with the most dispatches in the selected year.
         */
        topDispatchedVehicles: Array<{
            /**
             * Vehicle internal number.
             */
            internalNumber: string;
            /**
             * Station name assigned to the vehicle.
             */
            stationName: string;
            /**
             * Number of dispatches.
             */
            count: number;
        }>;
    };
};

export type GetYearRecapResponse = GetYearRecapResponses[keyof GetYearRecapResponses];